pyproject.toml
# poetry self add poetry-bumpversion

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool]

[tool.black]
line-length = 120
target-version = ['py39']

[tool.ruff]
select = ["ALL"]
ignore = []
line-length = 120

[tool.ruff.mccabe]
max-complexity = 18

[tool.isort]
profile = "black"

[tool.mypy]
pretty = true
python_version = "3.9"
show_error_codes = true
strict = true

[tool.poetry]
authors = ["vircoys <vircoys@gmail.com>"]
maintainers = ["eggplants <w10776e8w@yahoo.co.jp>"]
classifiers = [
  "License :: OSI Approved :: The Unlicense (Unlicense)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only"
]
description = "Online comic downloader"
keywords = ["downloader", "comic", "manga"]
name = "ccdl"
packages = [{include = "ccdl"}]
license = "Unlicense"
readme = "README.md"
repository = "https://github.com/eggplants/ccdl"
version = "0.3.1"

[tool.poetry.dependencies]
python = ">=3.8,<4"
pillow = "^9.4.0"
aiohttp = "^3.8.4"
lxml = "^4.9.2"
numpy = "^1.24.2"
pyexecjs = "^1.5.1"
requests = "^2.28.2"
selenium = "^4.8.0"
urllib3 = "^1.26.14"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
pre-commit = "^2.20.0"
taskipy = "^1.10.3"
aiohttp = "^3.8.4"
lxml-stubs = "^0.4.0"
numpy = "^1.24.2"
types-pillow = "^9.4.0.12"
types-urllib3 = "^1.26.25.6"
types-requests = "^2.28.11.13"
types-selenium = "^3.141.9"

[tool.poetry.scripts]
ccdl = "ccdl.main:main"

[tool.poetry_bumpversion.file."ccdl/__init__.py"]

[tool.taskipy.tasks]
lint = "pre-commit run -a"
profile = "python -m cProfile"
